actions/utils.test.ts
import * as utils from './utils'
import {exec} from '@actions/exec'
import * as core from '@actions/core'

jest.mock('@actions/core')
jest.mock('@actions/exec')

// just making sure the mock methods are correctly typed
const mockedExec = exec as jest.MockedFunction<typeof exec>

// reset the counter on mock fn calls after every test
afterEach(() => jest.clearAllMocks())

describe(`Actions Utils`, () => {
    test(`isHeadAncestor uses git CLI to check if the commit is part of the current branch, returns true when it is`, async () => {
        mockedExec.mockResolvedValue(0)
        const hash = '5265ef99f1c8e18bcd282a11a4b752731cad5665'
        const output = await utils.isHeadAncestor(hash)
        expect(mockedExec).toHaveBeenCalledWith('git merge-base', ['--is-ancestor', hash, 'HEAD'])
        expect(output).toBe(true)
    })

    test(`isHeadAncestor uses git CLI to check if the commit is part of the current branch, returns false when it is not`, async () => {
        mockedExec.mockRejectedValue(128)
        const hash = '5265ef99f1c8e18bcd282a11a4b752731cad5665'
        const output = await utils.isHeadAncestor(hash)
        expect(mockedExec).toHaveBeenCalledWith('git merge-base', ['--is-ancestor', hash, 'HEAD'])
        expect(output).toBe(false)
    })

    test(`getTreeHashForCommitHash uses git CLI to check if the commit is part of the current branch, returns false when it is not`, async () => {
        mockedExec.mockResolvedValue(0)
        const hash = '5265ef99f1c8e18bcd282a11a4b752731cad5665'
        const output = await utils.getTreeHashForCommitHash(hash)
        expect(mockedExec).toHaveBeenCalledWith(
            'git rev-parse',
            ['5265ef99f1c8e18bcd282a11a4b752731cad5665:'],
            {
                listeners: {stdout: expect.any(Function)}
            }
        )
        expect(output).toBe('')
    })

    test(`getCurrentRepoTreeHash uses git CLI to return the latest tree hash of the root of the repo`, async () => {
        mockedExec.mockResolvedValue(0)
        const output = await utils.getCurrentRepoTreeHash()
        expect(mockedExec).toHaveBeenCalledWith('git rev-parse', ['HEAD:'], {
            listeners: {stdout: expect.any(Function)}
        })
        expect(output).toBe('')
    })

    test(`writeLineToFile creates a file using a shell script`, async () => {
        mockedExec.mockResolvedValue(0)
        await utils.writeLineToFile({path: '/some/file', text: 'hello world'})
        expect(mockedExec).toHaveBeenCalledWith(`/bin/bash -c "echo hello world > /some/file"`)
    })

    describe('S3 Utils', () => {
        test(`fileExistsInS3 uses AWS CLI to check for of an object in S3 bucket, returns true if it exists`, async () => {
            mockedExec.mockResolvedValue(0)
            const output = await utils.fileExistsInS3({key: 'my/key', bucket: 'my-bucket'})
            expect(mockedExec).toHaveBeenCalledWith('aws s3api head-object', [
                '--bucket=my-bucket',
                '--key=my/key'
            ])
            expect(output).toBe(true)
        })

        test(`fileExistsInS3 uses AWS CLI to check for of an object in S3 bucket, returns true if it exists`, async () => {
            mockedExec.mockRejectedValue(255)
            const output = await utils.fileExistsInS3({key: 'my/key', bucket: 'my-bucket'})
            expect(mockedExec).toHaveBeenCalledWith('aws s3api head-object', [
                '--bucket=my-bucket',
                '--key=my/key'
            ])
            expect(output).toBe(false)
        })

        test(`copyFileToS3 uses AWS CLI to copy a local file to S3 bucket`, async () => {
            mockedExec.mockResolvedValue(0)
            await utils.copyFileToS3({path: '/some/file', key: 'my/key', bucket: 'my-bucket'})
            expect(mockedExec).toHaveBeenCalledWith('aws s3 cp', [
                '/some/file',
                's3://my-bucket/my/key'
            ])
        })

        test(`removeFileFromS3 uses AWS CLI to delete a file from S3 bucket`, async () => {
            mockedExec.mockResolvedValue(0)
            await utils.removeFileFromS3({key: 'my/key', bucket: 'my-bucket'})
            expect(mockedExec).toHaveBeenCalledWith('aws s3 rm', ['s3://my-bucket/my/key'])
        })
    })

    describe('Branch Sanitize - getSanitizedBranchName', () => {
        test(`sanitizes a full git ref into a DNS-ready string`, async () => {
            const output = utils.getSanitizedBranchName('refs/heads/hello/world')
            expect(output).toBe('hello-world')
        })

        test(`replaces all non-word characters with a dash`, async () => {
            const output = utils.getSanitizedBranchName(
                'refs/heads/hello/world-100%_ready,for.this!here:it"a)b(c{d}e'
            )
            expect(output).toBe('hello-world-100-_ready-for-this-here-it-a-b-c-d-e')
        })

        test(`removes multiple dashes in a row`, async () => {
            const output = utils.getSanitizedBranchName(
                'refs/heads/hello/my-very-weird_branch-100%%%-original'
            )
            expect(output).toBe('hello-my-very-weird_branch-100-original')
        })

        test(`caps the length of the sanitized name to 60 characters`, async () => {
            const output = utils.getSanitizedBranchName(
                'refs/heads/hello/my-very-weird_branch-100-original-whoooohoooooo-lets-do_it'
            )
            expect(output).toBe('hello-my-very-weird_branch-100-original-whoooohoooooo-lets-d')
        })
    })
})
